Renamed API clients used for e2e tests
