+++ test/e2e_tests/backend/inbucketClient.e2e.ts
@@ -19,7 +19,7 @@
 
 import axios, {AxiosInstance} from 'axios';
 
-export class InbucketClient {
+export class InbucketClientE2E {
   private readonly axiosInstance: AxiosInstance;
   private readonly inbucketUsername;
   private readonly inbucketPassword;
